cleaning up some stuff

+++ assignment-one/base-program/linear.matrix.multiply.c
@@ -4,7 +4,7 @@
 #include <stdbool.h>
 #include "common.h"
 
-void multiply(int size, int mat1[][size], int mat2[][size], int res[][size])
+void multiply(int size, double mat1[][size], double mat2[][size], double res[][size])
 {
     int i, j, k;
     for (i = 0; i < size; i++)
@@ -21,14 +21,14 @@ void multiply(int size, int mat1[][size], int mat2[][size], int res[][size])
 double run_linear_matrix_multiply(int size, bool print)
 {
     // Define the Arrays
-    int left_a[size][size];
-    int right_a[size][size];
-    int result_a[size][size];
+    double left_a[size][size];
+    double right_a[size][size];
+    double result_a[size][size];
 
     // Fill the Arrays
     create_two_d_array(size, size, left_a);
     create_two_d_array(size, size, right_a);
-    
+
     clock_t start = clock();
     multiply(size, left_a, right_a, result_a);
     clock_t end = clock();
